+++ OsmAnd/src/net/osmand/plus/views/layers/MapControlsLayer.java
@@ -374,8 +374,7 @@ public class MapControlsLayer extends OsmandMapLayer {
 			mapButton.setShowBottomButtons(showBottomMenuButtons);
 			mapButton.update();
 		}
-		mapActivity.updateFlockFreeHudControls(!isRoutePlanningMode && !isRouteFollowingMode
-				&& !isRouteDialogOpened && visibilityHelper.shouldShowTopButtons());
+		mapActivity.updateFlockFreeHudControls(!isRouteDialogOpened && visibilityHelper.shouldShowTopButtons());
 	}
 
 	@Override